Update app.py
The main gist of https://github.com/pmaji/crypto-whale-watching-app/pull/31 which wasn't merged was to include the total price of a bubble, which I think is valuable information (https://i.imgur.com/8ZwXOIL.png).

I also changed the TICKERS tuple into a list, because if someone just wants to use for example ETH-USD and removes the rest, Python will consider it as a string and not a tuple, so it will return an error.

I changed n_unique_orders from float into int because this is always a full number, and saying there are e.g. "5.0" unique orders instead of "5" is redundant.

I also improved some text formatting and added rounding that @theimo1221 forgot in the ladder whale tooltips (it would say e.g. 3.000000005345 BTC).

Also changed the server IP from 0.0.0.0 to 127.0.0.1